+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 + Added: options to enable/disable frames.
 + Added: command /xct help.
++ Changed: disabling combat text extra/others events messages.
 
 ## Changes in 2.0-release