+++ docs/CHANGELOG.md
@@ -4,8 +4,10 @@
 
 + Added: options to enable/disable frames.
 + Added: command /xct help.
++ Fixed: damage/healing done missed message.
 + Changed: disabling combat text extra/others events messages.
 + Changed: low health/mana message color.
++ Changed: disable scrollable frames by default.
 
 ## Changes in 2.0-release